MPI: revert rev 80 as it breaks on bethe
it turns out that OpenMPI produces a different set of libraries depending on
what language dialects the compiler understands making prediction (ie in CST
stage before the actual configure runs) what the libraries wil be quite
difficult. It is also impossible right now to overwrite which libraries will be
used when the code is build from scratch.


git-svn-id: http://svn.cactuscode.org/projects/ExternalLibraries/MPI/trunk@81 043a8217-7a68-40fe-abfd-36aa7d4fa6a8